organize fred.rs module
